+++ oss/client.go
@@ -6,7 +6,6 @@ import (
 	"bytes"
 	"encoding/xml"
 	"fmt"
-	"io"
 	"io/ioutil"
 	"log"
 	"net"
@@ -20,13 +19,6 @@ import (
 // Users use oss.New to create Client instance.
 //
 type (
-	// Client OSS client
-	Client struct {
-		Config     *Config      // OSS client configuration
-		Conn       *Conn        // Send HTTP request
-		HTTPClient *http.Client //http.Client to use - if nil will make its own
-	}
-
 	// ClientOption client option such as UseCname, Timeout, SecurityToken.
 	ClientOption func(*Client)
 )
@@ -92,8 +84,8 @@ func (client Client) Bucket(bucketName string) (*Bucket, error) {
 	}
 
 	return &Bucket{
-		client,
-		bucketName,
+		Client: client,
+		BucketName: bucketName,
 	}, nil
 }
 
@@ -1970,42 +1962,4 @@ func InsecureSkipVerify(enabled bool) ClientOption {
 	return func(client *Client) {
 		client.Config.InsecureSkipVerify = enabled
 	}
-}
-
-// Private
-func (client Client) do(method, bucketName string, params map[string]interface{},
-	headers map[string]string, data io.Reader, options ...Option) (*Response, error) {
-	err := CheckBucketName(bucketName)
-	if len(bucketName) > 0 && err != nil {
-		return nil, err
-	}
-
-	// option headers
-	addHeaders := make(map[string]string)
-	err = handleOptions(addHeaders, options)
-	if err != nil {
-		return nil, err
-	}
-
-	// merge header
-	if headers == nil {
-		headers = make(map[string]string)
-	}
-
-	for k, v := range addHeaders {
-		if _, ok := headers[k]; !ok {
-			headers[k] = v
-		}
-	}
-
-	resp, err := client.Conn.Do(method, bucketName, "", params, headers, data, 0, nil)
-
-	// get response header
-	respHeader, _ := FindOption(options, responseHeader, nil)
-	if respHeader != nil {
-		pRespHeader := respHeader.(*http.Header)
-		*pRespHeader = resp.Headers
-	}
-
-	return resp, err
 }
