+++ oss/conn.go
@@ -20,13 +20,6 @@ import (
 	"time"
 )
 
-// Conn defines OSS Conn
-type Conn struct {
-	config *Config
-	url    *urlMaker
-	client *http.Client
-}
-
 var signKeyList = []string{"acl", "uploads", "location", "cors",
 	"logging", "website", "referer", "lifecycle",
 	"delete", "append", "tagging", "objectMeta",
@@ -85,6 +78,7 @@ func (conn *Conn) init(config *Config, urlMaker *urlMaker, client *http.Client) 
 	return nil
 }
 
+
 // Do sends request and returns the response
 func (conn Conn) Do(method, bucketName, objectName string, params map[string]interface{}, headers map[string]string,
 	data io.Reader, initCRC uint64, listener ProgressListener) (*Response, error) {
@@ -93,79 +87,6 @@ func (conn Conn) Do(method, bucketName, objectName string, params map[string]int
 	uri := conn.url.getURL(bucketName, objectName, urlParams)
 	resource := conn.getResource(bucketName, objectName, subResource)
 	return conn.doRequest(method, uri, resource, headers, data, initCRC, listener)
-}
-
-// DoURL sends the request with signed URL and returns the response result.
-func (conn Conn) DoURL(method HTTPMethod, signedURL string, headers map[string]string,
-	data io.Reader, initCRC uint64, listener ProgressListener) (*Response, error) {
-	// Get URI from signedURL
-	uri, err := url.ParseRequestURI(signedURL)
-	if err != nil {
-		return nil, err
-	}
-
-	m := strings.ToUpper(string(method))
-	req := &http.Request{
-		Method:     m,
-		URL:        uri,
-		Proto:      "HTTP/1.1",
-		ProtoMajor: 1,
-		ProtoMinor: 1,
-		Header:     make(http.Header),
-		Host:       uri.Host,
-	}
-
-	tracker := &readerTracker{completedBytes: 0}
-	fd, crc := conn.handleBody(req, data, initCRC, listener, tracker)
-	if fd != nil {
-		defer func() {
-			fd.Close()
-			os.Remove(fd.Name())
-		}()
-	}
-
-	if conn.config.IsAuthProxy {
-		auth := conn.config.ProxyUser + ":" + conn.config.ProxyPassword
-		basic := "Basic " + base64.StdEncoding.EncodeToString([]byte(auth))
-		req.Header.Set("Proxy-Authorization", basic)
-	}
-
-	req.Header.Set(HTTPHeaderHost, req.Host)
-	req.Header.Set(HTTPHeaderUserAgent, conn.config.UserAgent)
-
-	if headers != nil {
-		for k, v := range headers {
-			req.Header.Set(k, v)
-		}
-	}
-
-	// Transfer started
-	event := newProgressEvent(TransferStartedEvent, 0, req.ContentLength, 0)
-	publishProgress(listener, event)
-
-	if conn.config.LogLevel >= Debug {
-		conn.LoggerHTTPReq(req)
-	}
-
-	resp, err := conn.client.Do(req)
-	if err != nil {
-		// Transfer failed
-		event = newProgressEvent(TransferFailedEvent, tracker.completedBytes, req.ContentLength, 0)
-		publishProgress(listener, event)
-		conn.config.WriteLog(Debug, "[Resp:%p]http error:%s\n", req, err.Error())
-		return nil, err
-	}
-
-	if conn.config.LogLevel >= Debug {
-		//print out http resp
-		conn.LoggerHTTPResp(req, resp)
-	}
-
-	// Transfer completed
-	event = newProgressEvent(TransferCompletedEvent, tracker.completedBytes, req.ContentLength, 0)
-	publishProgress(listener, event)
-
-	return conn.handleResponse(resp, crc)
 }
 
 func (conn Conn) getURLParams(params map[string]interface{}) string {
@@ -253,81 +174,7 @@ func (conn Conn) getResource(bucketName, objectName, subResource string) string 
 	return fmt.Sprintf("/%s/%s%s", bucketName, objectName, subResource)
 }
 
-func (conn Conn) doRequest(method string, uri *url.URL, canonicalizedResource string, headers map[string]string,
-	data io.Reader, initCRC uint64, listener ProgressListener) (*Response, error) {
-	method = strings.ToUpper(method)
-	req := &http.Request{
-		Method:     method,
-		URL:        uri,
-		Proto:      "HTTP/1.1",
-		ProtoMajor: 1,
-		ProtoMinor: 1,
-		Header:     make(http.Header),
-		Host:       uri.Host,
-	}
 
-	tracker := &readerTracker{completedBytes: 0}
-	fd, crc := conn.handleBody(req, data, initCRC, listener, tracker)
-	if fd != nil {
-		defer func() {
-			fd.Close()
-			os.Remove(fd.Name())
-		}()
-	}
-
-	if conn.config.IsAuthProxy {
-		auth := conn.config.ProxyUser + ":" + conn.config.ProxyPassword
-		basic := "Basic " + base64.StdEncoding.EncodeToString([]byte(auth))
-		req.Header.Set("Proxy-Authorization", basic)
-	}
-
-	date := time.Now().UTC().Format(http.TimeFormat)
-	req.Header.Set(HTTPHeaderDate, date)
-	req.Header.Set(HTTPHeaderHost, req.Host)
-	req.Header.Set(HTTPHeaderUserAgent, conn.config.UserAgent)
-
-	akIf := conn.config.GetCredentials()
-	if akIf.GetSecurityToken() != "" {
-		req.Header.Set(HTTPHeaderOssSecurityToken, akIf.GetSecurityToken())
-	}
-
-	if headers != nil {
-		for k, v := range headers {
-			req.Header.Set(k, v)
-		}
-	}
-
-	conn.signHeader(req, canonicalizedResource)
-
-	// Transfer started
-	event := newProgressEvent(TransferStartedEvent, 0, req.ContentLength, 0)
-	publishProgress(listener, event)
-
-	if conn.config.LogLevel >= Debug {
-		conn.LoggerHTTPReq(req)
-	}
-
-	resp, err := conn.client.Do(req)
-
-	if err != nil {
-		// Transfer failed
-		event = newProgressEvent(TransferFailedEvent, tracker.completedBytes, req.ContentLength, 0)
-		publishProgress(listener, event)
-		conn.config.WriteLog(Debug, "[Resp:%p]http error:%s\n", req, err.Error())
-		return nil, err
-	}
-
-	if conn.config.LogLevel >= Debug {
-		//print out http resp
-		conn.LoggerHTTPResp(req, resp)
-	}
-
-	// Transfer completed
-	event = newProgressEvent(TransferCompletedEvent, tracker.completedBytes, req.ContentLength, 0)
-	publishProgress(listener, event)
-
-	return conn.handleResponse(resp, crc)
-}
 
 func (conn Conn) signURL(method HTTPMethod, bucketName, objectName string, expiration int64, params map[string]interface{}, headers map[string]string) string {
 	akIf := conn.config.GetCredentials()
